+++ CHANGELOG.md
@@ -7,6 +7,7 @@ v1.4.0 plans to add a View and Model object, and several other minor additions a
 ### Additions
 
 - [Issue #77] - DatabaseColumn unique
+- [Issue #76] - Create View module
 - [Issue #73] - DatabaseColumn should allow array in the constructor
 - [Issue #72] - Create Model module